Update MAMP instruction link to better answer
I came up with a better solution for using MAMP’s `php` and `mysql` executables from the command line and posted an answer with it in the same Stack Overflow question that this page links to. The new solution by default exports the latest MAMP PHP version, rather than hardcoding the version number, so it’s more likely to allow people to actually get WP-CLI working in a MAMP environment.